requirements.txt
arrow==1.3.0
click==8.1.8
loguru==0.7.3
tabulate==0.9.0
python-dotenv==1.0.1
sqlmodel==0.0.23